Make the code usable on OSX.
Make the code compilable on as many platforms as possible. Expect the dead code
elimination to work to remove unreferenced private functions. Compiling them
makes it easier to do changes and ensure they are at least valid Go code yet
without having to run them on irrelevant platforms. Since Go compilation is so
fast, it's not a significantl cost to compile code that is discarded at link
time compared to the benefit of ensuring that the code is at least valid.