Remove PinOut.Set() and merge it into PinOut.Out().
The rationale is that when a pin is set as output, it must have an initial value
otherwise this creates a glitch.

Rewrite a lot of the code to be more resilient.

Improve pioselftest to test more thoroughly. Add automatic detection of sysfs
and optional enforcement of sysfs.

Fix Edges() to not be a leak anymore.

Rename sysfs.GetPin() to sysfs.PinByNumber().

Move IR_* detection from bcm283x (oops) to pins.