Rename drivers*.go to pio*.go, merge doc.go into pio.go.
This is to reduce the number of top level files, so users have to scroll less
when reading the README.md.

Move the Philosophy section lower in the README.md, since it's long and verbose
and users have to skip it to get to the Users section, which may be to far for
some.