Delete most of pio, intentionally break users by renaming it.
This code base will go away soon so people shouldn't depend on it.

Rename the directory to remove any potential confusion.

Some comments were not updated but it doesn't matter as the end goal is to
delete this code.

If this commit breaks you, please pin your checkout to the previous commit
271b43995600bf900389c8349da85a4a3771ce34 for now. Thanks and sorry for the
inconvenience.